+++ ErsatzTV.Core/Scheduling/PlayoutModeSchedulerDuration.cs
@@ -165,7 +165,7 @@ public class PlayoutModeSchedulerDuration(ILogger logger)
                     SubtitleMode = scheduleItem.SubtitleMode,
                     PlayoutItemWatermarks = [],
                     PlayoutItemGraphicsElements = [],
-                    SchedulingContext = GetSchedulingContext(scheduleItem, contentEnumerator)
+                    SchedulingContext = GetSchedulingContext(scheduleItem, null, contentEnumerator)
                 };
 
                 foreach (ProgramScheduleItemWatermark programScheduleItemWatermark in scheduleItem