+++ ErsatzTV.Core/Scheduling/PlayoutModeSchedulerMultiple.cs
@@ -108,7 +108,7 @@ public class PlayoutModeSchedulerMultiple(Map<CollectionKey, int> collectionItem
                 SubtitleMode = scheduleItem.SubtitleMode,
                 PlayoutItemWatermarks = [],
                 PlayoutItemGraphicsElements = [],
-                SchedulingContext = GetSchedulingContext(scheduleItem, contentEnumerator)
+                SchedulingContext = GetSchedulingContext(scheduleItem, null, contentEnumerator)
             };
 
             foreach (ProgramScheduleItemWatermark programScheduleItemWatermark in scheduleItem